Settings
Finished audio settings, added text for the other types that im going to fill in later on. Made the Hp stam and mana into bars. Fixed some minor bugs and bad code

+++ 2D Game/Assets/Scripts/Bullet.cs
@@ -7,7 +7,7 @@ public class Bullet : MonoBehaviour
 
     public Rigidbody2D rb;
 
-    [SerializeField] private int bulletDmg = 20;
+    [SerializeField] private float bulletDmg = 20f;
 
     public GameObject impactEffect;
